Implement ConnectionAdapter module to prepare for http support.
The ConnectionAdapter module encapsulates connections under a
WebSocket like interface. Ideally I will be able to mimick the
interface with http long polling and a javascript client that
stores the open connection's ID and references it when posting
new events to the server. When receiving the post request, the
server should retrieve the open connection using it's ID and
fire the #onmessage event with the newly received message.

I'm not sure this is the best way to implement this or that it
will even work reliably.