Add UserMnaager for sending to specific users from anywhere in your app.
The user manager is accessible through the WebsocketRails.user method.
The key for the hash is defined in `config.user_identifier` and will
be called on the `current_user` object if one exists at the time the
connection is established. It defaults to connection.id if there is
no `current_user` object accessible or the `current_user` object does
not respond to the identifier.