add window._pong_data to make pong data available
Allow to send extra data to the pong event to track active connections.  Use case: to use with an expiring key in Redis for active connection and on the pong, use the info in the _pong_data to extend the key.  This way we are no longer depending on the connected/disconnected event to keep track of active status but have a far more reliable way via Redis.